#143 change location of filter_device_list function
